+++ examples/nextjs-notification-settings/data/users.ts
@@ -1,0 +1,40 @@
+import type { User } from "@/types/data";
+
+export const users: User[] = [
+  // {
+  //   id: "[YOUR EMAIL ADDRESS]",
+  //   name: "[YOUR DISPlAY NAME]",
+  //   color: "[YOUR COLOR]",
+  //   picture: "https://liveblocks.io/avatars/avatar-0.png",
+  // },
+  {
+    id: "charlie.layne@example.com",
+    name: "Charlie Layne",
+    color: "#D583F0",
+    picture: "https://liveblocks.io/avatars/avatar-1.png",
+  },
+  {
+    id: "mislav.abha@example.com",
+    name: "Mislav Abha",
+    color: "#F08385",
+    picture: "https://liveblocks.io/avatars/avatar-2.png",
+  },
+  {
+    id: "tatum.paolo@example.com",
+    name: "Tatum Paolo",
+    color: "#F0D885",
+    picture: "https://liveblocks.io/avatars/avatar-3.png",
+  },
+  {
+    id: "anjali.wanda@example.com",
+    name: "Anjali Wanda",
+    color: "#85EED6",
+    picture: "https://liveblocks.io/avatars/avatar-4.png",
+  },
+  {
+    id: "emil.joyce@example.com",
+    name: "Emil Joyce",
+    color: "#8594F0",
+    picture: "https://liveblocks.io/avatars/avatar-6.png",
+  },
+];